Update using-dev-build
This updates the `using-dev-build` tests:
* The test runner is ported to typescript for maintainability
* chromedriver is updated to 104
* Formatter is run
* dfx.json is updated to work with dfx 0.11
