+++ demos/using-dev-build/package.json
@@ -4,11 +4,11 @@
   "private": true,
   "scripts": {
     "wdio": "wdio run ./wdio.conf.ts",
-    "test": "./test.js",
+    "test": "./test.ts",
     "format": "prettier --write --plugin-search-dir=. .",
     "build": "webpack",
     "start": "webpack serve --no-client-overlay",
-    "proxy": "./test.js --no-run"
+    "proxy": "./test.ts --no-run"
   },
   "devDependencies": {
     "@dfinity/auth-client": "^0.9.3",
@@ -16,7 +16,7 @@
     "@wdio/local-runner": "7.16.13",
     "@wdio/mocha-framework": "7.16.13",
     "@wdio/spec-reporter": "7.16.13",
-    "chromedriver": "git+https://github.com/Beltarm/node-chromedriver.git#chromedriver-102",
+    "chromedriver": "104.0.0",
     "copy-webpack-plugin": "^9.0.1",
     "html-webpack-plugin": "5.3.2",
     "prettier": "^2.5.0",
